refactor(html): move rendering helper modules under html rendering namespace
