+++ spec/lib/html2rss/html/rendering/description_builder_spec.rb
@@ -3,7 +3,7 @@
 require 'spec_helper'
 require 'nokogiri'
 
-RSpec.describe Html2rss::Rendering::DescriptionBuilder do
+RSpec.describe Html2rss::Html::Rendering::DescriptionBuilder do
   describe '#call' do
     context 'when base is plain text' do
       subject(:description) { described_class.new(base:, title: 'Sample instance', url: 'http://example.com', enclosures: nil, image: nil).call }
