+++ spec/lib/html2rss/html/rendering/video_renderer_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-RSpec.describe Html2rss::Rendering::VideoRenderer do
+RSpec.describe Html2rss::Html::Rendering::VideoRenderer do
   describe '#to_html' do
     it 'renders compact video html with escaped attributes', :aggregate_failures do # rubocop:disable RSpec/ExampleLength
       renderer = described_class.new(url: 'https://example.com/video.mp4?x=1&y=2', type: 'video/mp4')
